Prueba

+++ src/main/java/co/edu/escuelaing/cvds/ClothCraft/config/BasicAuthInterceptor.java
@@ -1,6 +1,5 @@
 package co.edu.escuelaing.cvds.ClothCraft.config;
 
-import jakarta.servlet.http.Cookie;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import lombok.extern.slf4j.Slf4j;
@@ -13,7 +12,6 @@ import co.edu.escuelaing.cvds.ClothCraft.repository.SessionRepository;
 
 import java.time.Duration;
 import java.time.Instant;
-import java.util.Arrays;
 import java.util.UUID;
 
 @SuppressWarnings("null")
@@ -27,39 +25,62 @@ public class BasicAuthInterceptor implements HandlerInterceptor {
         this.sessionRepository = sessionRepository;
     }
 
-
     private String getCookieValue(HttpServletRequest req, String cookieName) {
-        return Arrays.stream(req.getCookies())
-                .filter(c -> c.getName().equals(cookieName))
-                .findFirst()
-                .map(Cookie::getValue)
-                .orElse(null);
+        String cookieValue = req.getHeader(cookieName);
+        String sinAuthToken = null;
+        if(cookieValue != null) {
+            sinAuthToken = cookieValue.replace("authToken=", "");
+        }
+        else log.error("CookieValue is null");
+
+        return sinAuthToken;
     }
 
+    @SuppressWarnings("unused")
     @Override
-    public boolean preHandle(HttpServletRequest request, HttpServletResponse response, Object handler) throws Exception {
+    public boolean preHandle(HttpServletRequest request, HttpServletResponse response, Object handler)
+            throws Exception {
         log.info("BasicAuthInterceptor::preHandle()");
+        if ("OPTIONS".equals(request.getMethod())) {
+            log.info("Received OPTIONS request from origin: {}", request.getHeader("Origin"));
+            return true;
+        }
         String path = request.getRequestURI();
         log.info("Path:" + path);
-        String authToken = getCookieValue(request, "authToken");
+        String isStaticParam = request.getParameter("isStatic");
+        boolean isStatic = Boolean.parseBoolean(isStaticParam);
+        log.info("IsStatic: " + isStatic);
+        if (isStatic) {
+            return true;
+        }
+        String authToken = getCookieValue(request, "cookie");
         log.info("AuthToken: " + authToken);
         if (authToken != null) {
             Session session = sessionRepository.findByToken(UUID.fromString(authToken));
-            log.info("Session: " + session);
             if (session != null) {
+                log.info("Session: " + session.getToken() + " " + session.getUser().getEmail());
+                String userId = session.getUser().getId();
                 Duration duration = Duration.between(Instant.now(), session.getTimestamp());
                 long oneHour = 60L * 60L;
-                if(duration.getSeconds() > oneHour) {
+                if (duration.getSeconds() > oneHour) {
                     sessionRepository.delete(session);
                     response.sendError(HttpServletResponse.SC_EXPECTATION_FAILED, "SessionTimeout");
                     return false;
-                } 
-                /*else if (path.startsWith("/login/protected") && !session.getUser().getUserRoles().contains(UserRole.ADMINISTRADOR)) {
-                    response.sendError(HttpServletResponse.SC_FORBIDDEN, "Forbidden");
+                } else {
+
+                    String requestURI = request.getRequestURI();
+                    String queryString = request.getQueryString() != null ? request.getQueryString() : "";
+                    String userIdParam = "userId=" + userId;
+                    String newQueryString = queryString.isEmpty() ? userIdParam : queryString + "&" + userIdParam;
+
+                    // Agrega el parámetro isStatic=true al final de la cadena de consulta
+                    newQueryString += "&isStatic=true";
+
+                    String newRequestURI = requestURI + "?" + newQueryString;
+                    System.out.println("New Request URI: " + newRequestURI);
+                    // Actualizamos la URL de la solicitud
+                    request.getRequestDispatcher(newRequestURI).forward(request, response);
                     return false;
-                }*/ 
-                else {
-                    return true;
                 }
             } else {
                 response.sendError(HttpServletResponse.SC_BAD_REQUEST, "Bad Request");
@@ -71,13 +92,16 @@ public class BasicAuthInterceptor implements HandlerInterceptor {
         }
     }
 
+
     @Override
-    public void postHandle( HttpServletRequest request, HttpServletResponse response, Object handler, ModelAndView modelAndView) throws Exception {
+    public void postHandle(HttpServletRequest request, HttpServletResponse response, Object handler,
+            ModelAndView modelAndView) throws Exception {
         log.info("BasicAuthInterceptor::postHandle()");
     }
 
     @Override
-    public void afterCompletion(HttpServletRequest request, HttpServletResponse response, Object handler, Exception ex) throws Exception {
+    public void afterCompletion(HttpServletRequest request, HttpServletResponse response, Object handler, Exception ex)
+            throws Exception {
         log.info("BasicAuthInterceptor::afterCompletion()");
     }
 }
